add feature to generate cuesheets for playlists
The existing cuesheet feature allows cuesheets to be generated for
files, so you can have metadata of what is currently playing. This
change also adds the ability to export a cuesheet for the entire
playlist, so you can quickly jump from song to song, even if the
playlist extends into multiple files.