use uwsgi's attach-daemon to start celery
uWSGI actually has the ability to start daemons itself, so it makes
sense to take advantage of this. Instead of running separate celery
containers, we can just run them in the same container as the site
itself and avoid the additional overhead.

This reduces the need for supervisord, although it is still needed for
Dockerfile.dev for now.