Add "Show All DJs" functionality to Trackman
It's possible to mark a DJ as inactive. This is done with Automation by
default but can also be used to hide DJs who are inactive, i.e., haven't
played a show in say, the past 12 months.

Because DJs sometimes return, the ability to reuse old DJ information is
useful. This self-service "reactivation"  means we can more aggressively
run the pruning because it doesn't create any significant hassle for
DJs.

Fixes #240.