Sign session UUID when using Redis
Even though the session UUID isn't sensitive or predictable, it doesn't
hurt to sign it and it makes it impossible to attempt to brute force or
otherwise tamper with session UUIDs.

When deploying this, note that this will invalidate the current session
and any data stored in it, such as the CSRF token, which will cause 403
errors for anyone who happens to submit a form that was loaded before
the deployment. It might be a good idea to manually restart Trackman on
the workstations to avoid DJ confusion.