Individually autocomplete each field in Trackman
Previously, we were autocompleting based on the possible entries for
each field. This wasn't very intuitive for DJs and it made it
frustrating if a particular track didn't exist, but all the other
metadata did.

The solution is to autocomplete each field individually. We may still
have to make some tweaks to the sort order, but this should be a big
improvement.