Move radiothon mode variable to Redis
Having this in the config isn't really very helpful for a number of
reasons, including that changing it requires a restart of the site.
Instead, store the value of the variable in Redis. In the future, we can
consider adding a button or checkbox or something in the admin section
to enable/dsiable radiothon mode.