Refactor API and use new concept of being "on-air"
This change will allow DJs to login and start queuing tracks immediately
without going "on-air" until needed, allowing automation and other DJs
to continue uninterrupted. When they first play a track, the frontend
JavaScript will create a new DJSet if needed.

The queue is still only saved locally and cleared on logout, so this
won't allow DJs to queue up tracks on a different computer yet.