Move premium config to Redis
Since we often have to update premium configuration on short notice,
restarting the website to do so is problematic. This change allows
premium configuration to be changed without having to restart the site
and will also, in the future, allow for a "premium configuration"
section to be added to the admin area.