Renew DJ lease immediately up login (fix #316)
To avoid a situation where automation starts in between the time a DJ
logs in and the time the DJ creates their DJSet because of a short
`NO_DJ_TIMEOUT`, set the `dj_active` key to true and use the
`DJ_TIMEOUT` as the expiration.

Note that this extended automation start delay will persist if a DJ logs
in and immediately logs out without doing anything. This may or may not
be something we want to revisit in the future.